refactor(status): IncidentList splits into List + FilterChips + Card
The 372-line IncidentList becomes:
- IncidentList.svelte — shell: window filtering, monthly grouping,
  empty/loading placeholders, nowTick relative-time clock
- IncidentFilterChips.svelte — service chip row (uses FilterChip with
  underline variant) + roving-tabindex keyboard handler
- IncidentCard.svelte — single expandable incident article

The shell passes formatRelative down so the nowTick dependency lives
with the list, not per-card. Styling stays scoped to each component;
no visual change.
