+++ sites/status.arolariu.ro/src/lib/components/IncidentList.svelte
@@ -2,9 +2,8 @@
   import {onMount} from "svelte";
   import type {Incident, IncidentsFile, FilterWindow} from "../types/status";
   import {WINDOW_CONFIGS} from "../types/status";
-  import {formatDuration} from "../aggregation/formatDuration";
-  import IncidentDetail from "./IncidentDetail.svelte";
-  import FilterChip from "./FilterChip.svelte";
+  import IncidentCard from "./IncidentCard.svelte";
+  import IncidentFilterChips from "./IncidentFilterChips.svelte";
 
   interface Props {
     incidents: IncidentsFile | null;
@@ -94,60 +93,16 @@
     }
     return groups;
   });
-
-  function chipLabel(chip: string | null): string {
-    return chip === null ? "All" : chip;
-  }
-
-  function selectChip(chip: string | null) {
-    selectedService = chip;
-  }
-
-  function handleChipKeydown(event: KeyboardEvent, index: number) {
-    const total = allChips.length;
-    if (total <= 1) return;
-    let newIndex = index;
-    switch (event.key) {
-      case "ArrowRight":
-      case "ArrowDown":
-        newIndex = (index + 1) % total;
-        break;
-      case "ArrowLeft":
-      case "ArrowUp":
-        newIndex = (index - 1 + total) % total;
-        break;
-      case "Home":
-        newIndex = 0;
-        break;
-      case "End":
-        newIndex = total - 1;
-        break;
-      default:
-        return; // Let Space/Enter bubble naturally
-    }
-    event.preventDefault();
-    selectChip(allChips[newIndex] ?? null);
-    // Focus the newly-selected chip for keyboard users.
-    const container = (event.currentTarget as HTMLElement).parentElement;
-    const target = container?.children[newIndex] as HTMLElement | undefined;
-    target?.focus();
-  }
 </script>
 
 <div class="incidents">
   <h2 class="heading">Recent incidents · last {windowFilter}</h2>
   {#if incidents && serviceChips.length > 0}
-    <div class="filter-chips" role="radiogroup" aria-label="Filter incidents by service">
-      {#each allChips as chip, i (chip ?? "__all__")}
-        <FilterChip
-          label={chipLabel(chip)}
-          active={selectedService === chip}
-          variant="underline"
-          onClick={() => selectChip(chip)}
-          onKeydown={(e) => handleChipKeydown(e, i)}
-        />
-      {/each}
-    </div>
+    <IncidentFilterChips
+      chips={allChips}
+      selected={selectedService}
+      onSelect={(chip) => (selectedService = chip)}
+    />
   {/if}
   {#if !incidents}
     <div class="placeholder">Loading incidents…</div>
@@ -157,41 +112,12 @@
     {#each grouped as group (group.label)}
       <h3 class="month-header">{group.label}</h3>
       {#each group.items as inc (inc.id)}
-        <article class="item" data-status={inc.status}>
-          <button
-            type="button"
-            class="item-head"
-            aria-expanded={expandedId === inc.id}
-            aria-controls={`incident-detail-${inc.id}`}
-            onclick={() => (expandedId = expandedId === inc.id ? null : inc.id)}
-          >
-            <div class="item-rail" aria-hidden="true"></div>
-            <div class="item-body">
-              <div class="head">
-                <strong class="service-name">{inc.service}{inc.subCheck ? ` · ${inc.subCheck}` : ""}</strong>
-                <span class="pill pill-{inc.status}">{inc.status === "open" ? "Ongoing" : "Resolved"}</span>
-                <span class="severity severity-{inc.severity.toLowerCase()}">{inc.severity}</span>
-                <span class="chevron" aria-hidden="true">{expandedId === inc.id ? "▾" : "▸"}</span>
-              </div>
-              <div class="meta" title={inc.startedAt}>
-                <span>Started {formatRelative(inc.startedAt)}</span>
-                {#if inc.status === "resolved"}
-                  <span class="sep">·</span>
-                  <span>lasted {formatDuration(inc.durationMs)}</span>
-                {/if}
-                {#if inc.reason}
-                  <span class="sep">·</span>
-                  <code>{inc.reason}</code>
-                {/if}
-              </div>
-            </div>
-          </button>
-          {#if expandedId === inc.id}
-            <div id={`incident-detail-${inc.id}`} class="item-detail-wrap">
-              <IncidentDetail incident={inc} />
-            </div>
-          {/if}
-        </article>
+        <IncidentCard
+          incident={inc}
+          expanded={expandedId === inc.id}
+          onToggle={() => (expandedId = expandedId === inc.id ? null : inc.id)}
+          {formatRelative}
+        />
       {/each}
     {/each}
   {/if}
@@ -214,14 +140,6 @@
     margin-right: 8px;
     font-weight: 600;
   }
-  .filter-chips {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 2px 4px;
-    margin-bottom: var(--sp-md);
-    padding-bottom: var(--sp-sm);
-    border-bottom: 1px solid var(--border);
-  }
   .month-header {
     font-family: var(--font-mono);
     font-weight: 400;
@@ -243,98 +161,5 @@
     background: var(--surface);
     border-radius: var(--radius-md);
     border: 1px solid var(--border);
-  }
-  .item {
-    background: transparent;
-    border: 0;
-    border-bottom: 1px solid var(--border);
-    border-radius: 0;
-    margin-bottom: 0;
-    overflow: hidden;
-    position: relative;
-  }
-  .item:last-child { border-bottom: 0; }
-  .item-head {
-    all: unset;
-    display: grid;
-    grid-template-columns: 2px 1fr;
-    gap: var(--sp-md);
-    padding: var(--sp-md);
-    width: 100%;
-    box-sizing: border-box;
-    cursor: pointer;
-    color: var(--text);
-    transition: background 0.18s ease;
-  }
-  .item-head:focus-visible {
-    outline: 0;
-    background: var(--surface-hover);
-  }
-  .item-head:hover { background: var(--surface-hover); }
-  .item-rail {
-    background: var(--status-deg);
-    border-radius: 4px;
-  }
-  .item[data-status="open"] .item-rail { background: var(--status-down); }
-  .item-body { min-width: 0; }
-  .chevron {
-    margin-left: auto;
-    opacity: 0.5;
-    font-size: 11px;
-  }
-  .item-detail-wrap {
-    padding: 0 var(--sp-md) var(--sp-sm);
-  }
-  .head {
-    display: flex;
-    align-items: center;
-    flex-wrap: wrap;
-    gap: var(--sp-xs);
-    margin-bottom: 4px;
-  }
-  .service-name {
-    font-family: var(--font-mono);
-    font-weight: 500;
-    font-size: 12.5px;
-    letter-spacing: 0.005em;
-    color: var(--text);
-  }
-  .pill {
-    padding: 1px 6px;
-    border-radius: 2px;
-    font-family: var(--font-mono);
-    font-size: 9.5px;
-    font-weight: 500;
-    text-transform: uppercase;
-    letter-spacing: 0.1em;
-    background: transparent;
-    border: 1px solid currentColor;
-  }
-  .pill-open    { color: var(--status-down); }
-  .pill-resolved { color: var(--status-up); }
-  .severity {
-    font-family: var(--font-body);
-    font-size: 10px;
-    color: var(--text-muted);
-    text-transform: uppercase;
-    letter-spacing: 0.12em;
-    font-weight: 500;
-  }
-  .severity-unhealthy { color: var(--status-down); }
-  .severity-degraded { color: var(--status-deg); }
-  .meta {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    gap: var(--sp-2xs) var(--sp-xs);
-    font-size: var(--fs-xs);
-    opacity: 0.7;
-  }
-  .sep { opacity: 0.5; }
-  code {
-    background: var(--surface-hover);
-    padding: 1px 5px;
-    border-radius: 3px;
-    font-size: 10.5px;
   }
 </style>
